exp/knative-ml/ml/requirements.txt
parliament-functions==0.1.0
boto3==1.34.41